# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for SchemeSetu government scheme discovery and application tracking
 * OpenAPI spec version: 0.1.0
 */
import type { SavedScheme } from './savedScheme';

export interface DashboardSummary {
  profile_complete: boolean;
  total_matches: number;
  saved: number;
  applied: number;
  approved: number;
  rejected: number;
  recent_saved: SavedScheme[];
}