# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for SchemeSetu government scheme discovery and application tracking
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileInput } from './profileInput';

export type Profile = ProfileInput & {
  id: number;
  user_id: number;
  created_at: Date;
  updated_at: Date;
};