# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for SchemeSetu government scheme discovery and application tracking
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}